requirements.txt
albumentations==1.3.0
brotlipy==0.7.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
contourpy==1.0.5
cryptography @ file:///tmp/build/80754af9/cryptography_1652083738073/work
cycler==0.11.0
fonttools==4.37.4
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.22.1
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.1
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
networkx==2.8.7
numpy @ file:///tmp/abs_653_j00fmm/croots/recipe/numpy_and_numpy_base_1659432701727/work
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.5.0
Pillow==9.2.0
ply==3.11
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
pytz==2022.4
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.2
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
threadpoolctl==3.1.0
tifffile==2022.10.10
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tqdm==4.64.1
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3 @ file:///tmp/abs_5dhwnz6atv/croots/recipe/urllib3_1659110457909/work
